docs(link): provide playground with working controls
